requirements.txt
fastapi
uvicorn
pydantic
crewai
python-dotenv
openai